producer/requirements_producer.txt
kafka-python==2.0.2
Faker==19.6.2